Seperate partitionStats to another file
