+++ internal/storage/field_stats.go
@@ -305,55 +305,3 @@ func DeserializeFieldStatsList(blob *Blob) ([]*FieldStats, error) {
 	}
 	return stats, nil
 }
-
-type PartitionStatsSnapshot struct {
-	SegmentStats map[UniqueID][]FieldStats `json:"segmentStats"`
-	version      int64
-}
-
-func NewPartitionStatsSnapshot() *PartitionStatsSnapshot {
-	return &PartitionStatsSnapshot{
-		SegmentStats: make(map[UniqueID][]FieldStats, 0),
-	}
-}
-
-func (ps *PartitionStatsSnapshot) GetVersion() int64 {
-	return ps.version
-}
-
-func (ps *PartitionStatsSnapshot) SetVersion(v int64) {
-	ps.version = v
-}
-
-func (ps *PartitionStatsSnapshot) AddStats(segmentID UniqueID, stats []FieldStats) {
-	if _, ok := ps.SegmentStats[segmentID]; ok {
-		ps.SegmentStats[segmentID] = append(ps.SegmentStats[segmentID], stats...)
-	} else {
-		ps.SegmentStats[segmentID] = stats
-	}
-}
-
-func DeserializePartitionsStatsSnapshot(data []byte) (*PartitionStatsSnapshot, error) {
-	var messageMap map[string]*json.RawMessage
-	err := json.Unmarshal(data, &messageMap)
-	if err != nil {
-		return nil, err
-	}
-
-	partitionStats := &PartitionStatsSnapshot{
-		SegmentStats: make(map[UniqueID][]FieldStats),
-	}
-	err = json.Unmarshal(*messageMap["segmentStats"], &partitionStats.SegmentStats)
-	if err != nil {
-		return nil, err
-	}
-	return partitionStats, nil
-}
-
-func SerializePartitionStatsSnapshot(partStats *PartitionStatsSnapshot) ([]byte, error) {
-	partData, err := json.Marshal(partStats)
-	if err != nil {
-		return nil, err
-	}
-	return partData, nil
-}
