+++ internal/storage/field_stats_test.go
@@ -416,52 +416,3 @@ func TestMultiFieldStats(t *testing.T) {
 	assert.Equal(t, true, partitionKeyStats.Max.EQ(maxPk2))
 	assert.Equal(t, true, partitionKeyStats.Min.EQ(minPk2))
 }
-
-func TestPartitionStats(t *testing.T) {
-	partStats := NewPartitionStatsSnapshot()
-	{
-		fieldStats := make([]FieldStats, 0)
-		fieldStat1 := FieldStats{
-			FieldID: 1,
-			Type:    schemapb.DataType_Int64,
-			Max:     NewInt64FieldValue(200),
-			Min:     NewInt64FieldValue(100),
-		}
-		fieldStat2 := FieldStats{
-			FieldID: 2,
-			Type:    schemapb.DataType_Int64,
-			Max:     NewInt64FieldValue(200),
-			Min:     NewInt64FieldValue(100),
-		}
-		fieldStats = append(fieldStats, fieldStat1)
-		fieldStats = append(fieldStats, fieldStat2)
-		partStats.AddStats(1, fieldStats)
-	}
-	{
-		fieldStat1 := FieldStats{
-			FieldID: 1,
-			Type:    schemapb.DataType_Int64,
-			Max:     NewInt64FieldValue(200),
-			Min:     NewInt64FieldValue(100),
-		}
-		fieldStat2 := FieldStats{
-			FieldID: 2,
-			Type:    schemapb.DataType_Int64,
-			Max:     NewInt64FieldValue(200),
-			Min:     NewInt64FieldValue(100),
-		}
-		partStats.AddStats(2, []FieldStats{fieldStat1})
-		partStats.AddStats(2, []FieldStats{fieldStat2})
-	}
-	partStats.SetVersion(100)
-	assert.Equal(t, int64(100), partStats.GetVersion())
-	partBytes, err := SerializePartitionStatsSnapshot(partStats)
-	assert.NoError(t, err)
-	assert.NotNil(t, partBytes)
-	desPartStats, err := DeserializePartitionsStatsSnapshot(partBytes)
-	assert.NoError(t, err)
-	assert.NotNil(t, desPartStats)
-	assert.Equal(t, 2, len(desPartStats.SegmentStats))
-	assert.Equal(t, 2, len(desPartStats.SegmentStats[1]))
-	assert.Equal(t, 2, len(desPartStats.SegmentStats[2]))
-}
